* [Mail/Groups/Buckets] Group inbox buckets are now actual records in the database. Previously, inboxes were virtual records with id=0. This simplified some early workflows since 4.0, but it also made several queries less efficient because the group had to be compared every time as well.  For instance, filtering by a 'Sales' inbox would required filters like (group_id=3 and bucket_id=0) rather than just (bucket_id=5).  The 7.0 update automatically creates an inbox bucket for each group and moves the appropriate tickets into it, and the configuration of Virtual Attendant behaviors, mail routing rules, and worklists will be automatically updated.

+++ features/cerberusweb.core/api/uri/config/groups.php
@@ -94,24 +94,28 @@ class PageSection_SetupGroups extends Extension_PageSection {
 		
 		// If we're deleting the group
 		if(!empty($id) && !empty($delete)) {
-			if(!empty($delete_move_id)) {
-				if(null != ($group = DAO_Group::get($id))) {
-					$fields = array(
-						DAO_Ticket::GROUP_ID => $delete_move_id
-					);
-					$where = sprintf("%s=%d",
-						DAO_Ticket::GROUP_ID,
-						$id
-					);
-					DAO_Ticket::updateWhere($fields, $where);
-					
-					// If this was the default group, move it.
-					if($group->is_default)
-						DAO_Group::setDefaultGroup($delete_move_id);
-					
-					DAO_Group::delete($group->id);
-				}
+			// Verify our destination exists
+			if(
+				!empty($delete_move_id)
+				&& false != ($move_to_group = DAO_Group::get($delete_move_id))
+				&& false != ($move_to_bucket = $move_to_group->getDefaultBucket())
+				) {
+			
+				$fields = array(
+					DAO_Ticket::GROUP_ID => $move_to_group->id,
+					DAO_Ticket::BUCKET_ID => $move_to_bucket->id,
+				);
+				$where = sprintf("%s=%d",
+					DAO_Ticket::GROUP_ID,
+					$id
+				);
+				DAO_Ticket::updateWhere($fields, $where);
 				
+				// If this was the default group, move it.
+				if($group->is_default)
+					DAO_Group::setDefaultGroup($move_to_group->id);
+				
+				DAO_Group::delete($group->id);
 			}
 			
 		} elseif(!empty($id)) {
